refactor: extract now-playing IGDB import and add-platform helpers
Move promptNowPlayingAddPlatformSelection into now-playing/nowPlayingAddService.ts
and the IGDB import helpers (startNowPlayingIgdbImport,
startNowPlayingIgdbImportFromInteraction, importGameFromIgdb) into
now-playing/nowPlayingIgdbImport.service.ts. The IGDB service imports the
add-platform helper, keeping the dependency one-directional (no cycles).

Refs #894
